// ui/src/generated/rust-types/NovelContainerOutput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NovelSourceOutput } from "./NovelSourceOutput";

export type NovelContainerOutput = { id: string, name: string, type: string, icon: string | null, color: string | null, description: string | null, posterPath: string | null, scrapeEnabled: boolean, scrapeAgents: Array<string> | null, sortOrder: number, settings: unknown | null, syncStatus: string, lastSyncAt: string | null, itemCount: number, sources: Array<NovelSourceOutput>, createdAt: string, updatedAt: string, };